feat: give jest configurator tests to error view
