+++ packages/error-view/__tests__/android/error-view.android.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Article test on android", () => {
+  shared();
+});
