+++ packages/error-view/__tests__/ios/error-view.ios.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Article test on iOS", () => {
+  shared();
+});
