+++ packages/error-view/__tests__/web/error-view.web.test.js
@@ -1,0 +1,5 @@
+import shared from "../shared";
+
+describe("Article test on web", () => {
+  shared();
+});
